feat:Ensure that before deleting the offerteam ( by the leader of course ) ensure that it has no member

+++ controllers/teamOfferController.js
@@ -223,6 +223,9 @@ exports.deleteTeamOffer = async (req, res) => {
   try {
     const teamOffer = await prisma.teamOffer.findUnique({
       where: { id },
+      include: {
+        TeamMembers: true, // Include team members to check if there are any
+      },
     });
 
     if (!teamOffer) {
@@ -235,9 +238,16 @@ exports.deleteTeamOffer = async (req, res) => {
         .json({ error: "You are not the leader of this team offer" });
     }
 
-    await prisma.teamOffer.delete({
-      where: { id },
-    });
+    if (teamOffer.TeamMembers.length > 0) {
+      return res
+        .status(400)
+        .json({ error: "You cannot delete a team offer with members" });
+    }
+
+    if (teamOffer)
+      await prisma.teamOffer.delete({
+        where: { id },
+      });
 
     await prisma.student.update({
       where: { id: leader_id },
